feat: minimalist sh parser
Enhances ParserBuilder with Parser interface
