+++ internal/parser/shparser_test.go
@@ -1,0 +1,18 @@
+package parser
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestShParserWithContent(t *testing.T) {
+	shContent := `true\nfalse`
+
+	parser, _ := NewParserBuilder("sh").
+		WithContent(shContent).
+		Build()
+
+	commands := parser.Parse()
+	assert.Equal(t, 2, len(commands))
+}
